+++ libs/vedana-etl/src/vedana_etl/pipeline.py
@@ -7,23 +7,14 @@ from vedana_etl.catalog import (
     dm_anchors,
     dm_attributes,
     dm_links,
-    dm_version,
     edges,
     grist_edges,
-    grist_edges_filtered,
     grist_nodes,
-    grist_nodes_filtered,
-    llm_embeddings_config,
-    llm_pipeline_config,
     memgraph_edges,
     memgraph_indexes,
     memgraph_nodes,
     memgraph_vector_indexes,
     nodes,
-    judge_config,
-    eval_gds,
-    eval_llm_answers,
-    eval_llm_judge,
 )
 
 data_model_steps = [
@@ -31,12 +22,6 @@ data_model_steps = [
         func=steps.get_data_model,  # Generator with main graph data
         outputs=[dm_anchors, dm_attributes, dm_links],
         labels=[("flow", "regular"), ("flow", "on-demand"), ("stage", "extract"), ("stage", "data-model")],
-    ),
-    BatchGenerate(
-        func=steps.get_data_model_snapshot,
-        delete_stale=False,
-        outputs=[dm_version],
-        labels=[("flow", "eval"), ("flow", "regular"), ("flow", "on-demand"), ("stage", "data-model")],
     ),
 ]
 
@@ -46,20 +31,6 @@ grist_steps = [
         outputs=[grist_nodes, grist_edges],
         labels=[("flow", "on-demand"), ("stage", "extract"), ("stage", "grist")],
     ),
-    BatchTransform(
-        func=steps.filter_grist_nodes,
-        inputs=[grist_nodes, dm_anchors, dm_attributes],
-        outputs=[grist_nodes_filtered],
-        labels=[("flow", "on-demand"), ("stage", "transform"), ("stage", "grist")],
-        transform_keys=["node_id"],
-    ),
-    BatchTransform(
-        func=steps.filter_grist_edges,
-        inputs=[grist_edges, dm_links],
-        outputs=[grist_edges_filtered],
-        labels=[("flow", "on-demand"), ("stage", "transform"), ("stage", "grist")],
-        transform_keys=["from_node_id", "to_node_id", "edge_label"],
-    ),
 ]
 
 # ---
@@ -68,14 +39,14 @@ grist_steps = [
 default_custom_steps = [
     BatchTransform(
         func=steps.prepare_nodes,
-        inputs=[grist_nodes_filtered],
+        inputs=[grist_nodes],
         outputs=[nodes],
         labels=[("flow", "on-demand"), ("stage", "transform"), ("stage", "grist")],
         transform_keys=["node_id"],
     ),
     BatchTransform(
         func=steps.prepare_edges,
-        inputs=[grist_edges_filtered],
+        inputs=[grist_edges],
         outputs=[edges],
         labels=[("flow", "on-demand"), ("stage", "transform"), ("stage", "grist")],
         transform_keys=["from_node_id", "to_node_id", "edge_label"],